LayoutTests/fast/xpath/text-nodes-expected.txt

<doc><elem>a<![CDATA[b]]>c</elem></doc>
child::*, [object Element]: ""
child::node(), [object Element]: "a b c"
descendant::*, [object Element]: ""
descendant::node(), [object Element]: "a b c"
descendant::node()[2], [object Element]: "b"
ancestor-or-self::node(), b: "b null null null"
ancestor-or-self::*, b: "null null"
ancestor-or-self::node(), a: "a null null null"
ancestor-or-self::*, a: "null null"
following::node(), [object Element]: ""
following::node(), a: "b c"
following::text(), a: "b c"
following::node(), b: "c"
following-sibling::node(), [object Element]: ""
following-sibling::node(), a: "b c"
following-sibling::text(), a: "b c"
following-sibling::node(), b: "c"
preceding::node(), b: "a"
preceding-sibling::node(), b: "a"
preceding::node(), c: "b a"
preceding::text(), c: "b a"
preceding-sibling::node(), c: "b a"
preceding-sibling::text(), c: "b a"
self::node(), b: "b"
